requirements.txt
# Core dependencies
python-nostr>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0

# Type checking
mypy>=1.5.0
types-requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0